+++ src/app/types/search.parameter.ts
@@ -1,7 +1,10 @@
 import { SelectedObservationCodes } from './selected-observation-codes';
 
 export interface SearchParameter {
-  name?: string;
+  // Search parameter name for HTTP request
+  element?: string;
+  // Search parameter display name
+  displayName?: string;
   // TODO: value type TBD
   value?: any;
   selectedObservationCodes?: SelectedObservationCodes[];